Simulator view integration.
There are still some details. To do that I has to integrate a logging system.  With the logger some debug text areas are no longer needed in the views.

Simulator and parser are now separated. The simulator takes the intermediate representation of the parser. To do this I had to improve the parsing process.

Work on the memory view.